refactor(P2): 事件契约上移 src/shared/protocol.js，运行时与门禁同源
- 新增 src/shared/protocol.js 承载 AGENT_EVENT_TYPES(26) 与 INBOUND_SOCKET_EVENTS(40)，逐字保序
- scripts/agent-event-contract.js 改为 import 后再导出，既有 import 面不变
- agent.js 的 emit/emitTransient 增加 type 自检：未登记则 [event-contract] 告警但照常发出
  只覆盖经 AgentSession 的 17 型，其余 9 型走服务端广播路径仍由门禁静态扫描把关
- 新增 4 条单测（含全量扫契约防清单与校验分叉）
